create function interval and call from index
