+++ src/helpers/bot.ts
@@ -1,20 +1,45 @@
 import File from "./file";
+import { getTickerPair } from "./../services/ticker";
 const Bot = {
   calculateDiff: (initial: number, current: number): number => {
     return (100 * (initial - current)) / ((initial + current) / 2);
   },
-  writeLog: (diff: number): void => {
-    if (diff < -0.01) {
+  writeLog: (diff: number, percentDiff: number): void => {
+    if (diff < -percentDiff) {
       File.appendFile(
         "./logs.txt",
-        "[Decreasing] Ask < -0.01% on time: " + new Date().toLocaleString()
+        `[Decreasing] Ask < -${percentDiff}% on time: ${new Date().toLocaleString()}`
       );
-    } else if (diff > 0.01) {
+    } else if (diff > percentDiff) {
       File.appendFile(
         "./logs.txt",
-        "[Increasing] Ask > +0.01% on time: " + new Date().toLocaleString()
+        `[Increasing] Ask > +${percentDiff}% on time: ${new Date().toLocaleString()}`
       );
     }
+  },
+  interval: (
+    isFirstTime: boolean = true,
+    initialValue: number = 0,
+    intervalMilSec: number = 5000,
+    percentDiff: number,
+    pair: Array<string>
+  ): void => {
+    setInterval(async () => {
+      const { data } = await getTickerPair("BTC-USD");
+
+      if (isFirstTime && data) {
+        const initialValueString = JSON.parse(data).ask;
+        initialValue = +initialValueString;
+        isFirstTime = false;
+      } else if (!isFirstTime && data) {
+        const actualValueString: string = JSON.parse(data).ask;
+        const actualValue: number = +actualValueString;
+
+        const diff = Bot.calculateDiff(initialValue, actualValue);
+
+        Bot.writeLog(diff, percentDiff);
+      }
+    }, intervalMilSec);
   },
 };
