+++ src/index.ts
@@ -1,4 +1,4 @@
-import { getTickerPair } from "./services/ticker";
+import { getTickerPair, getTickerByCurrency } from "./services/ticker";
 import Bot from "./helpers/bot";
 
 let isFirstTime: boolean = true;
@@ -7,19 +7,21 @@ let isFirstTime: boolean = true;
 let initialValue: number = 0;
 
 // if API was a WebSocket or SSE i won't use a setInterval
-setInterval(async () => {
-  const { data } = await getTickerPair("BTC-USD");
+// setInterval(async () => {
+//   const { data } = await getTickerPair("BTC-USD");
 
-  if (isFirstTime && data) {
-    const initialValueString = JSON.parse(data).ask;
-    initialValue = +initialValueString;
-    isFirstTime = false;
-  } else if (!isFirstTime && data) {
-    const actualValueString: string = JSON.parse(data).ask;
-    const actualValue: number = +actualValueString;
+//   if (isFirstTime && data) {
+//     const initialValueString = JSON.parse(data).ask;
+//     initialValue = +initialValueString;
+//     isFirstTime = false;
+//   } else if (!isFirstTime && data) {
+//     const actualValueString: string = JSON.parse(data).ask;
+//     const actualValue: number = +actualValueString;
 
-    const diff = Bot.calculateDiff(initialValue, actualValue);
+//     const diff = Bot.calculateDiff(initialValue, actualValue);
 
-    Bot.writeLog(diff);
-  }
-}, 5000);
+//     Bot.writeLog(diff,0.01);
+//   }
+// }, 5000);
+
+Bot.interval(true, 0, 5000, 0.01, ["BTC-USD"]);
